+++ modules/core/__init__.py
@@ -1,0 +1,6 @@
+from modules.core.device_registry import DeviceRegistry
+from modules.core.plugin import Plugin
+from modules.core.plugin_manager import PluginManager
+from modules.core.widget_registry import WidgetRegistry
+
+__all__ = ["DeviceRegistry", "Plugin", "PluginManager", "WidgetRegistry"]